fix(core): split setAttributes for workflow VM bundle
CI build error: the umbrella workflow package re-exports
\`@workflow/core/_workflow\`, which (via my earlier setAttributes
export) transitively pulled \`step/context-storage\` and therefore
\`node:async_hooks\`. The workflow VM bundle's no-Node-module
constraint rejected it.

Split into three modules:
- set-attributes-shared.ts: validation + 'use step' dispatcher. No
  contextStorage import, safe in both bundles.
- set-attributes.ts (step/host): looks up runId via contextStorage,
  falls back to WORKFLOW_CONTEXT_SYMBOL. Re-exported from core/index.
- workflow/set-attributes.ts (VM): reads only WORKFLOW_CONTEXT_SYMBOL.
  Re-exported from core/_workflow.

Mirrors the same dual-context layout as getWorkflowMetadata.
